backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlmodel>=0.0.16
email-validator>=2.1
python-dotenv>=1.0
